move CreateTagModal to separate file
