+++ src/component/form/ColorInput.tsx
@@ -1,5 +1,5 @@
 import ColorPicker from "component/common/ColorPicker";
-import Modal from "component/modal/Modal";
+import { Modal } from "component/common/Modal";
 import { useModal } from "hook/modal";
 import { useState } from "react";
 import { useFormContext } from "react-hook-form";
