+++ src/page/settings/Tags.tsx
@@ -1,73 +1,27 @@
-import { yupResolver } from "@hookform/resolvers/yup";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { AxiosError } from "axios";
-import Button from "component/button/Button";
 import Heading from "component/common/Heading";
 import Tag from "component/common/Tag";
-import ColorInput from "component/form/ColorInput";
-import Input from "component/form/Input";
 import PageContent from "component/layout/PageContent";
 import PageLayout from "component/layout/PageLayout";
-import Modal from "component/modal/Modal";
+import { CreateTagModal } from "component/modal/CreateTagModal";
 import { TAGS_QUERY_KEY } from "constant/queries";
 import useTusksApi from "hook/api";
 import { useModal } from "hook/modal";
 import { useToast } from "provider/ToastProvider";
 import { FC } from "react";
-import { FormProvider, useForm } from "react-hook-form";
-import { INewTag, ITag } from "type";
-import { TAG_SCHEMA } from "validation";
+import { ITag } from "type";
 
 const Tags: FC = () => {
-  const { fetchAllTags, createNewTag, deleteTag } = useTusksApi();
+  const { fetchAllTags, deleteTag } = useTusksApi();
+  const { toast } = useToast();
+  const queryClient = useQueryClient();
   const { data, isLoading } = useQuery([TAGS_QUERY_KEY], fetchAllTags);
   const { isOpen, onClose, onOpen } = useModal();
-
-  const methods = useForm<INewTag>({
-    mode: "onChange",
-    resolver: yupResolver(TAG_SCHEMA),
-  });
-
-  const { toast } = useToast();
-  const {
-    handleSubmit,
-    register,
-    reset,
-    formState: { errors, isValid },
-  } = methods;
-  const queryClient = useQueryClient();
-
-  const { mutateAsync: updateAsync, isLoading: isUpdating } = useMutation<
-    ITag,
-    AxiosError,
-    INewTag
-  >((tag: INewTag) => createNewTag(tag));
 
   const { mutateAsync: deleteAsync } = useMutation<ITag, AxiosError, string>(
     (id: string) => deleteTag(id)
   );
-
-  const submit = async (tag: INewTag) => {
-    await updateAsync(tag, {
-      onSuccess: (data: ITag) => {
-        queryClient.setQueryData<ITag[]>([TAGS_QUERY_KEY], (original) => {
-          if (original) {
-            return [...original, data];
-          } else {
-            return [data];
-          }
-        });
-        onClose();
-      },
-      onError: (error) => {
-        toast({
-          icon: "warning",
-          title: "Hooops",
-          description: error.message,
-        });
-      },
-    });
-  };
 
   const onDelete = (id: string) => {
     deleteAsync(id, {
@@ -102,7 +56,6 @@ const Tags: FC = () => {
             className="rounded-full bg-primary-600 bg-opacity-20 px-3 py-2 text-primary-800 dark:bg-primary-200 dark:bg-opacity-10 dark:text-primary-300"
             onClick={(e) => {
               e.stopPropagation();
-              reset();
               onOpen();
             }}
           >
@@ -122,32 +75,7 @@ const Tags: FC = () => {
           {data?.length === 0 && <div>You have no tags, yet.</div>}
         </div>
       </PageContent>
-      <Modal isOpen={isOpen} onClose={onClose}>
-        <h1 className="text-lg font-black">Create tag</h1>
-        <FormProvider {...methods}>
-          <form
-            onSubmit={handleSubmit(submit)}
-            className="mt-4 flex flex-col gap-2"
-          >
-            <Input label="label" {...register("label")} error={errors?.label} />
-            <span className="mt-4">color</span>
-            <div className="flex flex-row gap-3">
-              <ColorInput label="light" name="color.light" />
-              <ColorInput label="dark" name="color.dark" />
-            </div>
-            <Button
-              icon="tag"
-              hoverIcon="check"
-              type="submit"
-              className="mt-4"
-              isSubmitting={isUpdating}
-              isDisabled={!isValid}
-            >
-              Create
-            </Button>
-          </form>
-        </FormProvider>
-      </Modal>
+      <CreateTagModal isOpen={isOpen} onClose={onClose} />
     </PageLayout>
   );
 };
